+++ apps/codeforafrica/src/lib/data/blockify/featuredProjects.js
@@ -1,16 +1,15 @@
-import { sortTags } from "@/codeforafrica/lib/data/utils/posts";
+import { sortTags } from "@/codeforafrica/lib/data/utils/tags";
 
-async function ourWorkShowcase(block) {
+async function featuredProjects(block) {
   const { projects } = block;
   const projectTags = projects.map(({ tag }) => tag).filter(Boolean);
   const tags = sortTags(projectTags);
 
   return {
-    tags,
     ...block,
-    projects,
+    tags,
     slug: block.blockType,
   };
 }
 
-export default ourWorkShowcase;
+export default featuredProjects;
